most of an internal search engine
This little baby is currently able to retrieve HTML content from any
file that's within its directory or any subdirectory. Right now, it just
displays the contents you specify but eventually I'm going to make it
send the retrieved content to a database to create an index for an
internal search engine for a website.