Massive Overhaul
There is still some work to be done, but I've come very far. I've
separated the functions out so they aren't dependent on one another
anymore. I've also added lots of exception handling. I designed the
load_content function to be flexible so that you can make alterations to
how much data you want scraped and that function will still work the
same, so long as all the input arrays have the same length and the DB is
prepared to handle any changes in data input. Next, I plan on making the
scrape_files function more flexible so that you can just specify an
array of what content to scrape off each file and it will adjust without
having to really hardcode anything in. I also want to make a blacklist
for the scan_directory function so that you can supply it with an array
of paths to files/folders to ignore and it will do so. That way it
doesn't do silly things like request itself or scrape files that you
don't care about, etc. I've also thought about trying to convert this
little library into a class. We'll see. And then I've thought about
adding some more features like a separate class/function that overwrites
the contents of the DB instead of just adding to it. We'll see!