Improved scrape_content()
scrape_content() now accepts an array of DOMXPath queries as a parameter
along with the original list of paths. This system is much more
flexible. Unfortunately, however, load_content() is not fond of the new
system so I have to make major alterations to the prepare_sql() and
prepare_data() func.'s in order to appease load_content().